+++ cookbooks/chub-buyspace/recipes/buyspace.rb
@@ -46,7 +46,7 @@ end
 directory "#{node['chub-buyspace']['config_dir']}" do
 	owner "tomcat7"
 	group "tomcat7"
-	mode 0770
+	mode 0775
 end
 
 directory "#{node['chub-buyspace']['app_dir']}" do
@@ -64,7 +64,7 @@ end
 directory "#{node['chub-buyspace']['log_dir']}" do
 	owner "tomcat7"
 	group "tomcat7"
-	mode 0770
+	mode 0775
 end
 
 directory "/etc/chadmin" do
@@ -88,12 +88,12 @@ image_dir_existed = File.exists?(image_dir)
 directory image_dir do
   owner "tomcat7"
   group "tomcat7"
-  mode 0770
+  mode 0777
   not_if {image_dir_existed}
 end
 
 execute "mount_windows_share" do
-  command "echo #{node['chub-buyspace']['images']['shareDirectory']} #{node['chub-buyspace']['images']['shareMount']} cifs username=#{node['chub-buyspace']['images']['shareUser']},pass=#{node['chub-buyspace']['images']['sharePassword']},iocharset=utf8 >> /etc/fstab"
+  command "echo #{node['chub-buyspace']['images']['shareDirectory']} #{node['chub-buyspace']['images']['shareMount']} cifs username=#{node['chub-buyspace']['images']['shareUser']},pass=#{node['chub-buyspace']['images']['sharePassword']},iocharset=utf8,rw,noperm >> /etc/fstab"
   not_if {image_dir_existed}
 end