+++ roles/missioncontrol_team.rb
@@ -7,13 +7,4 @@ default_attributes(
       }
   }
 )
-override_attributes(
-  "ssh_keys" => {
-    "lzarou" => "lzarou",
-    "wgrunenwald" => "wgrunenwald",
-    "kstahl" => "kstahl",
-    "bcarr" => "bcarr",
-    "dcross" => "dcross"
-  }
-)
 run_list "recipe[chub_mc_app::sensu_client]"