+++ roles/onboardingws_dev.rb
@@ -3,18 +3,7 @@ description "A linux machine running onboarding WS (in development)"
 run_list "recipe[base]", "recipe[chub_mc_webservices]", "recipe[chub_mc_webservices::onboarding_webservice]"
 default_attributes "chub_mc_webservice" => {
 	"onboarding" => {
-    	"jar_file_name" => "wsonb.jar",
-		"jar_file_url" => "overrideME"
+    	"jar_file_name" => "OnboardingService-standalone.jar",
+		"jar_file_url" => "http://bamboom1:8085/browse/MC-WSON/latest/artifact/shared/Onboarding-Web-Service-Jar/OnboardingService-standalone.jar?os_username=mess&os_password=messuser"
 		},
-		"config" => {
-			"salesforceClient" => {
-				"host" => "https://test.salesforce.com",
-				"clientId" => "",
-				"clientSecret" => "",
-				"username" => "",
-				"password" => ""
-
-			},
-		
-	},
 }